+++ Main.java
@@ -1,4 +1,4 @@
-import src.screen.MainScreen;
+import src.view.MainScreen;
 import src.theme.ThemeColor;
 
 import javax.swing.*;
@@ -6,7 +6,7 @@ import javax.swing.*;
 public class Main {
     public static void main(String[] args) {
         SwingUtilities.invokeLater(() -> {
-            ThemeColor.applyThemeDark("dark-ui.properties");
+            ThemeColor.applyThemeDark("deep-forest.properties");
             MainScreen mainScreen = new MainScreen();
             mainScreen.run();
         });
