+++ src/infrastructure/RecipeRepository.java
@@ -1,0 +1,9 @@
+package src.infrastructure;
+
+import src.entity.Recipe;
+
+import java.util.List;
+
+public interface RecipeRepository {
+    List<Recipe> loadAll();
+}
