--- src/repository/RecipeRepository.java
@@ -1,9 +1,0 @@
-package src.repository;
-
-import src.entity.Recipe;
-
-import java.util.List;
-
-public interface RecipeRepository {
-    List<Recipe> loadAll();
-}
