--- src/screen/navigation/NavigationHandler.java
@@ -1,8 +1,0 @@
-package src.screen.navigation;
-
-public interface NavigationHandler {
-    void displayHomeScreen();
-    void displayCategoryScreen();
-    void displayPlannerScreen();
-    void displayFavoriteScreen();
-}
