+++ src/view/navigation/NavigationHandler.java
@@ -1,0 +1,8 @@
+package src.view.navigation;
+
+public interface NavigationHandler {
+    void displayHomeScreen();
+    void displayCategoryScreen();
+    void displayPlannerScreen();
+    void displayFavoriteScreen();
+}
